# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample YourVibes backend server",
        "title": "API Documentation YourVibes backend",
        "termsOfService": "https://github.com/poin4003/yourVibes_GoApi",
        "contact": {
            "name": "TEAM HKTP",
            "url": "https://github.com/poin4003/yourVibes_GoApi",
            "email": "pchuy4003@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "yourvibesapi.duckdns.org:8080",
    "basePath": "/v1/2024",
    "paths": {
...
        "/advertise/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get many advertise",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertise_user"
                ],
                "summary": "Get many advertise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post_id to filter ads",
                        "name": "post_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit of ads per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "When user want to create advertise by post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "advertise_user"
                ],
                "summary": "Create advertise",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateAdvertiseRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/advertise/admin": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve advertise base on filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin_advertise_report"
                ],
                "summary": "Get a list of advertise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post id to get advertise",
                        "name": "post_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user email",
                        "name": "user_email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by from date",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by to date",
                        "name": "to_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by from price",
                        "name": "from_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by to price",
                        "name": "to_price",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Sort in descending order",
                        "name": "is_descending",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/advertise/{advertise_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve advertise",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin_advertise_report"
                ],
                "summary": "Get advertise detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Advertise ID",
                        "name": "advertise_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "request.CreateAdvertiseRequest": {
            "type": "object",
            "required": [
                "end_date",
                "post_id",
                "redirect_url",
                "start_date"
            ],
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "post_id": {
                    "type": "string"
                },
                "redirect_url": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token without 'Bearer ' prefix",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}